add full ts support and buildpipe; cleanup;
