+++ .gitignore
@@ -1,6 +1,7 @@
 # Build Folders (you can keep bin if you'd like, to store dlls and pdbs)
 src/[Bb]in/
 src/[Oo]bj/
+src/web/wwwroot/
 
 # Logs
 Logs/
