+++ src/core/Services/JWTGen.cs
@@ -17,12 +17,12 @@ public class JWTGen
         _jwtOptions = jwtOptions;
         _handler = new JwtSecurityTokenHandler();
     }
-    public string GenerateNewToken(string username)
+    public string GenerateNewToken(string username,string role)
     {
         var jwt = _jwtOptions.Value;
         var securityKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(jwt.key));
         var credentials = new SigningCredentials(securityKey, SecurityAlgorithms.HmacSha256);
-        IEnumerable<Claim> claims = [new Claim("name", username),new Claim("role", "manager")];
+        IEnumerable<Claim> claims = [new Claim("name", username),new Claim("role", role)];
         //if(username == "admin")
         //{
         //    claims.Append(new Claim("role", "manager"));
