+++ src/web/Controllers/ManagerController.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Mvc;
 using core.Services;
 
 
-[Authorize]
+[Authorize(Roles="manager")]
 public class ManagerController : Controller
 {
     IUserService _userService;
@@ -18,8 +18,11 @@ public class ManagerController : Controller
         _logService = logService;
         _userService = userService;
     }
-    public async Task<IActionResult> AddUser([FromBody]User user)
+    public async Task<IActionResult> AddUser([FromBody] User user)
     {
+        if (string.IsNullOrEmpty(user.Uname) ||
+           string.IsNullOrEmpty(user.Password))
+            return BadRequest("Some requiered fields are empty");
         _logger.LogInformation($"Add user request {user.Uname}");
         await _userService.NewUserAsync(user);
         await _userService.SaveChangesAsync();
@@ -35,5 +38,11 @@ public class ManagerController : Controller
         _logger.LogInformation($"Result of deleting user with id={id}:{result}");
 
         return result ? Ok() : BadRequest("Non existent user");
+    }
+    public IActionResult ManageUsers(){
+        return PartialView("/Views/Partials/_ManageUsers.cshtml",_userService.GetAll());
+    }
+    public async Task<IActionResult> ManageLogs(){
+        return PartialView("/Views/Partials/_ManageLogs.cshtml",_logService.GetAll());
     }
 }
