+++ src/web/Views/Pages/AddUser.cshtml
@@ -13,20 +13,23 @@
                     <h4 class="mb-0">Register New User</h4>
                 </div>
                 <div class="card-body">
-                    <form id="registerForm" name="signIn">
+                    <form id="registerForm" class="needs-validation" name="signIn" novalidate>
                         <div class="mb-3">
                             <label for="Uname" class="form-label">Username</label>
-                            <input type="text" class="form-control" id="fullName" name="uname" required>
+                            <input type="text" class="form-control" id="fullName" name="uname" pattern="^[a-zA-Z][a-zA-Z0-9_]{2,15}$" required>
+                            <div class="invalid-feedback">Bad username</div>
                         </div>
 
                         <div class="mb-3">
                             <label for="email" class="form-label">Email address</label>
-                            <input type="email" class="form-control" id="email" name="email" required>
+                            <input type="email" class="form-control" id="email" name="email">
+                            <div class="invalid-feedback">Bad email format!</div>
                         </div>
 
                         <div class="mb-3">
                             <label for="password" class="form-label">Password</label>
                             <input type="password" class="form-control" id="password" name="password" required>
+                            <div class="invalid-feedback">Password is required!</div>
                         </div>
 
                         <div class="mb-3">
@@ -36,6 +39,7 @@
                                 <option value="Admin">Admin</option>
                                 <option value="User">User</option>
                             </select>
+                            <div class="invalid-feedback">Role is required!</div>
                         </div>
 
                         <button type="submit" id="createBtn" class="btn btn-dark btn-outline-success w-100">Create
@@ -47,18 +51,23 @@
     </div>
 </div>
 @section Scripts{
-    <script type="module">
-        import * as service from '/js/manage/NewUser.js'
-        $('#registerForm').on('submit',async function(e){
-                e.preventDefault();
-                const form = document.forms.signIn;
-                const selected = form.elements.role.selectedIndex;
-                await service.NewUser(form.elements.uname.value,
-                        form.elements.password.value,
-                        form.elements.role.options[selected].value,
-                        form.elements.email.value);
+<script type="module">
+    import * as userService from '/js/manage/NewUser.js'
+    $('#registerForm').on('submit', async function (e) {
+        e.preventDefault();
+        e.stopPropagation();
+        const form = document.forms.signIn;
+        const uname = form.elements.uname.value;
+        if (form.checkValidity()) {
+            const selected = form.elements.role.selectedIndex;
+            await userService.NewUser(uname,
+                form.elements.password.value,
+                form.elements.role.options[selected].value,
+                form.elements.email.value);
+        }
+        form.classList.add("was-validated");
 
-        });
+    });
 
-    </script>
+</script>
 }
