+++ src/web/Views/Pages/Manage.cshtml
@@ -67,11 +67,11 @@ Layout = "_Layout";
                             <ul class="nav flex-column">
                                 <li class="nav-item"><a id="sidebar-item" class="nav-link active" aria-current="page"
                                         data-bs-dismiss="offcanvas" data-bs-target="#offcanvasResponsive"
-                                        href="/Pages/ManageUsers">User
+                                        asp-controller="Manager" asp-action="ManageUsers">User
                                         Base</a></li>
                                 <li class="nav-item"><a id="sidebar-item" class="nav-link" aria-current="page"
                                         data-bs-dismiss="offcanvas" data-bs-target="#offcanvasResponsive"
-                                        href="/Pages/ManageLogs">Logs</a>
+                                        asp-controller="Manager" asp-action="ManageLogs">Logs</a>
                                 </li>
                             </ul>
                         </div>
@@ -106,7 +106,7 @@ Layout = "_Layout";
     </div>
 </div>
 @section Scripts{
-    <script src="~/js/manageSideBar.js"></script>
+    <script src="~/js/manage/manageSideBar.js"></script>
     <script>
         $('.modal-footer .btn-secondary').click(function (e) {
             $('#modalConfirm') .modal('hide');
